# docs/api/openapi.yaml
openapi: 3.0.3
info:
  title: ダンプ運行記録システム API
  description: |
    ダンプトラック運行記録管理システムのRESTful API仕様書
    
    ## 機能概要
    - ユーザー認証・認可
    - 車両管理
    - 運行記録管理
    - GPS位置情報追跡
    - 統計・レポート機能
    
  version: 1.0.0
  contact:
    name: システム管理者
    email: admin@dumptracker.local
  license:
    name: MIT
    url: https://opensource.org/licenses/MIT
servers:
  - url: http://localhost:8000/api/v1
    description: 開発環境
  - url: http://10.1.119.244:8000/api/v1
    description: ステージング環境
security:
  - bearerAuth: []
paths:
...
  /users:
    get:
      tags:
        - ユーザー管理
      summary: ユーザー一覧取得
      description: ページネーション付きでユーザー一覧を取得（管理者・マネージャーのみ）
      parameters:
        - name: page
          in: query
          description: ページ番号
          schema:
            type: integer
            default: 1
        - name: limit
          in: query
          description: 1ページあたりの件数
          schema:
            type: integer
            default: 10
        - name: role
          in: query
          description: 役割フィルタ
          schema:
            type: string
            enum: [ADMIN, MANAGER, DRIVER]
        - name: search
          in: query
          description: 検索キーワード（名前、ユーザー名、メール）
          schema:
            type: string
      responses:
        '200':
          description: ユーザー一覧取得成功
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                    example: true
                  data:
                    type: object
                    properties:
                      users:
                        type: array
                        items:
                          $ref: '#/components/schemas/User'
                      pagination:
                        $ref: '#/components/schemas/Pagination'
        '403':
          $ref: '#/components/responses/Forbidden'
  /users/{id}:
    get:
      tags:
        - ユーザー管理
      summary: ユーザー詳細取得
      description: 指定されたIDのユーザー詳細情報を取得
      parameters:
        - name: id
          in: path
          required: true
          description: ユーザーID
          schema:
            type: string
            format: uuid
      responses:
        '200':
          description: ユーザー詳細取得成功
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                    example: true
                  data:
                    type: object
                    properties:
                      user:
                        $ref: '#/components/schemas/UserDetail'
        '404':
          $ref: '#/components/responses/NotFound'

  # 車両管理エンドポイント
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  schemas:
    User:
      type: object
      properties:
        id:
          type: string
          format: uuid
          description: ユーザーID
        username:
          type: string
          description: ユーザー名
        email:
          type: string
          format: email
          description: メールアドレス
        name:
          type: string
          description: 表示名
        role:
          type: string
          enum: [ADMIN, MANAGER, DRIVER]
          description: 役割
        employeeId:
          type: string
          description: 従業員ID
        phone:
          type: string
          description: 電話番号
        isActive:
          type: boolean
          description: アクティブフラグ
        lastLoginAt:
          type: string
          format: date-time
          description: 最終ログイン日時
        createdAt:
          type: string
          format: date-time
          description: 作成日時
        updatedAt:
          type: string
          format: date-time
          description: 更新日時
    UserDetail:
      allOf:
        - $ref: '#/components/schemas/User'
        - type: object
          properties:
            statistics:
              type: object
              properties:
                totalOperations:
                  type: integer
                  description: 総運行回数
                completedOperations:
                  type: integer
                  description: 完了した運行回数
                totalDistance:
                  type: number
                  description: 総走行距離
                lastOperationDate:
                  type: string
                  format: date
                  description: 最終運行日
...
    Pagination:
      type: object
      properties:
        currentPage:
          type: integer
          description: 現在のページ
        totalPages:
          type: integer
          description: 総ページ数
        totalCount:
          type: integer
          description: 総件数
        limit:
          type: integer
          description: 1ページあたりの件数
    Error:
      type: object
      properties:
        success:
          type: boolean
          example: false
        message:
          type: string
          description: エラーメッセージ
        error:
          type: string
          description: エラーコード
  responses:
...
    Forbidden:
      description: 権限エラー
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
          example:
            success: false
            message: "権限が不足しています"
            error: "FORBIDDEN"
    NotFound:
      description: リソースが見つからない
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
          example:
            success: false
            message: "リソースが見つかりません"
            error: "NOT_FOUND"
...